Большая переработка интерфейса

+++ server/src/modules/draft/draft.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Draft } from './entities/draft.entity';
@@ -9,7 +13,7 @@ import { UpdateDraftInput } from './inputs/update-draft.input';
 export class DraftService {
   constructor(
     @InjectRepository(Draft)
-    private repository: Repository<Draft>
+    private repository: Repository<Draft>,
   ) {}
 
   create(userId: number, createDraftInput: CreateDraftInput) {
@@ -20,17 +24,36 @@ export class DraftService {
     return `This action returns all draft`;
   }
 
+  findUserDrafts(userId: number) {
+    return this.repository.find({
+      where: { user: { id: userId } },
+      order: { createdAt: 'DESC' },
+    });
+  }
+
   findOne(id: number) {
     return this.repository.findOne({ where: { id } });
   }
 
   update(updateDraftInput: UpdateDraftInput) {
-    const { id, ...draftData } = updateDraftInput;
-    this.repository.update(id, draftData);
-    return id;
+    const { draftId, ...draftData } = updateDraftInput;
+    this.repository.update({ id: draftId }, draftData);
+    return draftId;
   }
 
-  remove(id: number) {
-    return this.repository.delete(id);
+  async remove(id: number, userId: number) {
+    const find = await this.repository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
+
+    if (!find) throw new NotFoundException('Статья не найдена');
+
+    if (userId !== find.user.id)
+      throw new ForbiddenException('Статья принадлежит другому пользователю.');
+
+    this.repository.delete(id);
+
+    return find;
   }
 }
